Add files via upload
Local Scope:
When a variable or function is defined inside a function, it is said to be in the local scope of that function. 
Variables defined in the local scope are only accessible within that function and are destroyed once the function 
completes execution. Local variables have the highest priority in accessing over global variables of the same name.